+++ src/main/java/org/example/VendingMachine/VendingMachine.java
@@ -6,7 +6,7 @@ public interface VendingMachine {
 
     void initProducts(List<Product> list);
 
-    Product getProduct(String name, double volume, double temp);
+    Product getProduct(String name, double volume, int temp);
 
     Product getProduct(String name, double volume);
 }
